+++ HISTORY.rst
@@ -5,6 +5,9 @@ History
 1.4.1 (unreleased)
 ------------------
 
+- Make bind user and password optional.
+  [thet]
+
 - Add ``ignore_cert`` option to ignore TLS/SSL certificate errors for self
   signed certificates when using the ``ldaps`` uri schema. See:
   http://stackoverflow.com/a/7810308/1337474